pyproject.toml
[project]
name = "YouTubeTranscriber"
version = "0.1.0"
description = "Extracts availble subtitles or generates them if they don't exist"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper",
    "yt-dlp",
    "youtube-transcript-api",
    "mcp"
]
